backend/requirements.txt
django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
langchain==0.1.9
langchain-openai==0.0.7
langgraph==0.0.19
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.4.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.28.1
